Add two paddles
The two paddles are controlled by two players. Player1 using w and s whereas player2 uses up and down arrows. The paddles stop and leave a small space away from the edge of the screen.

Keeps track of which buttons have been currently pressed inside an array called inputKeys in the GAME_VARIABLES variable. The status of these keys being pressed are only set to false on keyup event.

Each paddle is given the controls to know which key to check for before moving.