Rework ball rebound angle
Previously ball rebounded at fixed angle. Now calculates an angle up to 60 degrees. The ball is rebounded at a larger angle the farther away it is from the center of the paddle. The resulting direction is stored as a unit vector in direction property of ball.

The ball now uses this direction vector to keep track of which direction it is moving. And a speed property is used to dictate how fast the ball is moving.

Also, after a collision is done, the ball turns off collisions temporarily to avoid multiple hits on the same paddle, which caused it to be stuck inside the paddle.