Modify paddle sections
Changed the length of paddle sections. The center portion of the paddle will be length of 2 * sectionSize. The upper and lower sections of the paddle will get 3 * sectionSize. This should give the paddle sides more room and less space in the center. This reduces the amount of deflects at 90 degrees.