Modify ball speed based on speed of paddle
Adds extra speed to the ball when paddle is moving fast.

Changed the collision detection to cover a wider range by taking the back face of the previous position. This lets the ball collide even when coming from a sharp angle. But make sure ball is coming from the correct direction before allowing collision.