Add score function and game state
Added three game states, PREPARATION, PLAYING, and GAMEOVER.

Changed scoring to update the ScoreManager's playerScores. ArenaManager sets the gameState to GAMEOVER when score of 11 is reached.

Game begins with GAMEOVER state. Only ball is rendered, which bounces around the screen.

After 5 seconds, the state is changed to PREPARATION. The paddles positions and scoreBoard are reset before the next step.

After waiting another 2 seconds, the ball is spawned with a random direction. The ball is served to the player who lost the point last. Initially the ball is served to player 2.