Add a menu screen
Renamed ArenaManager to GameScreen, to be alongside the new MenuScreen. A ScreenManager is used to swap between the two.

A MenuScreen draws the Play button. Use an event listener on the canvas to check for clicks. Functions that find where the mouse position is and checks for intersection is stored in a Util object. This Util object is used by the MenuScreen to see if the player clicked the button.

Renamed the GameScreen updateArena and drawArena functions to just update and draw respectively. This is so that ScreenManager can call the same function depending on whether GameScreen or MenuScreen is currently selected.

The ScreenManager passes a reference to itself to the GameScreen and MenuScreen. This is so that the screens are able to tell the ScreenManager to change the current screen when, for example the player presses play or the game is over.

Renamed gameOver to preGame and postGame. Both play the same animation for 5 seconds but change to a different screen state.

Both screens will have a begin function when that screen is swapped to. This replaces the call to gameOver when the gameLoop starts.