Implement mode select
Add a Mode class stored in GAME_VARIABLES. New modes can be added by providing a name, a type of ball, and a description, which will be displayed on the ModeSelectScreen.

Add a new ModeSelectScreen class. The class retrieves information from Mode. The arrow buttons allow the player to cycle between modes added to the Mode class.

When the player starts a game, the GameScreen class also retrieves the ball from the selected mode. Removed ball argument in the GameScreen constructor.
Prepended missing 'this' calls to variables in the GameScreen class, as it was causing undefined ball errors when the ball was changed.

Add a drawCenteredAt function to ball, so that the ModeSelectScreen can draw the ball centered at a specific position.

Shift Button to be a parent class, so as to differentiate between buttons with text and buttons that are arrows.

Modify ScreenManager to have an addScreen function to add an arbitrary number of screens.

Add update, draw and begin functions to the Screen class. This handles drawing of all buttons associated with a screen, and provides default definitions so that children of Screen do not need to reimplement update and begin if no functionality is added.

Create new Ball for test cases, as original ball was removed.