Implement pause functionality
Use a checkPaused function which switches the gameState to PAUSED when the player presses the spacebar and the gameState is PLAYING or PREPARATION. The PAUSED state will draw the overlay to blur the background and display text to tell the player that the game is paused.
Prevent multiple presses by waiting until spacebar is released before accepting another pause.

Game used setTimeout to delay between states of the game. This prevented pausing whilst waiting for a setTimeout function. Implement delta time to calculate actual time elapsed into timer variable during PREPARATION state, and be able to stop calculating when game is in PAUSED state.

TODO: Allow rest of game to make use of delta time.