Add sounds
Play audio when the ball bounces on the bounds of the screen, bounces of a paddle, or when scoring.

Does not play sound for pre and post game. Added a canPlayAudio property to GAME_VARIABLES, which sets canPlayAudio to true after preGame is called and false before postGame is called.